Error handling and finished third layer
Added lots of error handling and finished the last layer for the API

+++ Videogadon/Videogadon/Controllers/CommentsController.cs
@@ -22,7 +22,6 @@ namespace Videogadon.Controllers
     {
         private readonly IGameCategoriesRepository _gameCategoriesRepository;
         private readonly IGamesRepository _gamesRepository;
-
         private readonly ICommentsRepository _commentsRepository;
 
         public CommentsController(IGameCategoriesRepository gameCategoriesRepository, IGamesRepository gamesRepository, ICommentsRepository commentsRepository)
@@ -33,10 +32,10 @@ namespace Videogadon.Controllers
         }
 
         [HttpGet]
-        public async Task<IEnumerable<CommentDto>> GetMany(int gameCategoryId, int gameId)
+        public async Task<IEnumerable<CommentDto>> GetMany()
         {
-            var comments = await _commentsRepository.GetAsync(gameCategoryId, gameId);
-            return comments.Select(x => new CommentDto(x.Id, x.Content, x.CreationDate));
+            var comments = await _commentsRepository.GetAsync();
+            return comments.Select(x => new CommentDto(x.Id, x.Content, x.CreationDate, x.GameId));
         }
 
         // api/gameCategories/{gameCategoryId}/games/{gameId}/comments/{commentId}
@@ -50,10 +49,10 @@ namespace Videogadon.Controllers
             var game = await _gamesRepository.GetAsync(gameCategoryId, gameId);
             if (game == null) return NotFound($"Couldn't find a game with id of {gameId}"); //404
 
-            var comment = await _commentsRepository.GetAsync(gameCategoryId, gameId, commentId);
-            if (comment == null) return NotFound($"Couldn't find a comment with id of {commentId}");
+            var comment = await _commentsRepository.GetAsync(gameId, commentId);
+            if (comment == null) return NotFound($"Couldn't find a comment with id of {commentId}"); //404
 
-            return new CommentDto(comment.Id, comment.Content, comment.CreationDate);
+            return new CommentDto(comment.Id, comment.Content, comment.CreationDate, comment.GameId);
         }
 
         // api/gameCategories/{gameCategoryId}/games/{gameId}/comments
@@ -63,7 +62,6 @@ namespace Videogadon.Controllers
             var gameCategory = await _gameCategoriesRepository.GetAsync(gameCategoryId);
             if (gameCategory == null)
             {
-                //404
                 return NotFound($"Couldn't find a game category with id of {gameCategoryId}"); //404
             }
 
@@ -78,12 +76,11 @@ namespace Videogadon.Controllers
                 Content = createCommentDto.Content,
                 CreationDate = DateTime.Now
             };
-            comment.GameCategoryId = gameCategoryId;
             comment.GameId = gameId;
             await _commentsRepository.CreateAsync(comment);
 
             //201
-            return Created($"api/gameCategories/{gameCategoryId}/games/{gameId}/comments/{comment.Id}", new CommentDto(comment.Id, comment.Content, comment.CreationDate));
+            return Created($"api/gameCategories/{gameCategoryId}/games/{gameId}/comments/{comment.Id}", new CommentDto(comment.Id, comment.Content, comment.CreationDate, comment.GameId));
         }
 
         // api/gameCategories/{gameCategoryId}/games/{gameId}/comments/{commentId}
@@ -100,10 +97,10 @@ namespace Videogadon.Controllers
             var game = await _gamesRepository.GetAsync(gameCategoryId, gameId);
             if (game == null)
             {
-                return NotFound($"Couldn't find a game category with id of {gameId}"); //404
+                return NotFound($"Couldn't find a game with id of {gameId}"); //404
             }
 
-            var comment = await _commentsRepository.GetAsync(gameCategoryId, gameId, commentId);
+            var comment = await _commentsRepository.GetAsync(gameId, commentId);
             if (comment == null)
             {
                 return NotFound($"Couldn't find a comment with id of {commentId}"); //404
@@ -112,22 +109,34 @@ namespace Videogadon.Controllers
             comment.Content = commentDto.Content;
             await _commentsRepository.UpdateAsync(comment);
 
-            return Ok(new CommentDto(comment.Id, comment.Content, comment.CreationDate));
+            return Ok(new CommentDto(comment.Id, comment.Content, comment.CreationDate, comment.GameId)); //200
         }
 
         // api/gameCategories/{gameCategoryId}/games/{gameId}/comments/{commentId}
         [HttpDelete]
         [Route("{commentId}")]
-        public async Task<ActionResult> Remove(int gameCategoryId, int gameId, int commentId)
+        public async Task<ActionResult> Remove(int gameCategoryId,int gameId, int commentId)
         {
-            var comment = await _commentsRepository.GetAsync(gameCategoryId, gameId, commentId);
+            var gameCategory = await _gameCategoriesRepository.GetAsync(gameCategoryId);
+            if (gameCategory == null)
+            {
+                return NotFound($"Couldn't find a game category with id of {gameCategoryId}"); //404
+            }
+
+            var game = await _gamesRepository.GetAsync(gameCategoryId, gameId);
+            if (game == null)
+            {
+                return NotFound($"Couldn't find a game with id of {gameId}"); //404
+            }
+
+            var comment = await _commentsRepository.GetAsync( gameId, commentId);
             if (comment == null)
             {
                 return NotFound($"Couldn't find a comment with id of {commentId}"); //404
             }
             await _commentsRepository.DeleteAsync(comment);
 
-            return NoContent();
+            return NoContent();// 204
         }
     }
 }
